% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrameProcessing.R
\name{candsetmaxpro}
\alias{candsetmaxpro}
\title{Create a maximum projection (MaxPro) design}
\usage{
candsetmaxpro(candmat, npoints = NULL, groupid = NULL)
}
\arguments{
\item{candmat}{matrix - A matrix of candidate points in the design space. Ideally should be a random uniform fill of the design space constructed using \code{\link{convexuniformfill}}}

\item{npoints}{integer - The number of points to use in the MaxPro design}

\item{groupid}{vector - group membership ID for all points in candmat}
}
\value{
\item{DesignMat}{A matrix with the resulting MaxPro design}
  \item{ObjFunc}{The value of the objective function being minimized to find the MaxPro design}
}
\description{
Function to find maximum projection design from set of candidate points and cluster IDs. Either npoints or groupid must be provided.
  TODO: ADD IN LOSS FUNCITON THAT USES nloptim (same approach and gradient function as MaxPro) AND DISTANCE FROM HULL AS ADDITIVE LOSS WITH EXTRA WEIGHT ON DISTANCE FROM HULL TO CREATE NUMERICAL OPTIMZIER THAT WILL WORK FOR MAXPRO
  DOUBLE CHECK LOSS FUNCTION. ALSO SEEMS TO HAVE ISSUE WITH GETTING STUCK IN LOCAL MINIMA SO WILL NOT WORK VERY WELL WITHOUT CLUSTERING FIRST.
}
