% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAnalysis.R
\name{choicesim}
\alias{choicesim}
\title{Simulate discrete choice results for a design using a specified true model}
\usage{
choicesim(testbase, trueform, trueparams, altvect, numsims,
  fullrandomize = FALSE, tournament = FALSE, randtournament = FALSE)
}
\arguments{
\item{testbase}{data frame - design matrix of the discrete choice design to use for simulation}

\item{trueform}{formula - formula used to construct model matrix for simulation}

\item{trueparams}{vector - vector of true coefficient values corresponding to the exact order of terms returned by a model.matrix call using trueform}

\item{altvect}{vector - vector of integers that assign each row in testbase to a choice set}

\item{numsims}{integer - number of repeats to use for simulation}

\item{fullrandomize}{logical - whether the order of profiles in testbase should be fully randomized before each simulation run
(i.e. randomly construct choice sets by sampling without replacement from the rows of testbase for each simulation run)}

\item{tournament}{logical - whether a single elimination winners' tournament should be used on the results of each simulation.}

\item{randtournament}{logical - whether the tournament bracket used for the tournament should be randomized (even if the design is not randomized per fullrandomize)}
}
\value{
\item{simframe}{data frame containing coefficient estimates resulting from running a model against each simulation in numsims}
  \item{optframe}{data frame containing estimated optimal parameter value combination to maximize trueform based on estimated coefficients in simframe}
}
\description{
This function simulates running a discrete choice study using numsims repeats and returns the results of the analysis.
  It is compatible with conventional and tournament discrete choice designs.
}
