% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrameProcessing.R
\name{convexuniformfill}
\alias{convexuniformfill}
\title{Fill a convex hull with a uniform random sample of points}
\usage{
convexuniformfill(n, inputranges, constrainnodes)
}
\arguments{
\item{n}{integer - number of random points to generate}

\item{inputranges}{matrix with named columns where first row is minimum and second row is maximum - range of base input variables that will be used as ranges to generate random points for design space filling (including geometric combination variables)}

\item{constrainnodes}{matrix with nodes of constraining manifold (hull) within rectangular inputranges space. Should be a matrix of nodes as produced by the convhulln geometry package. It is ok to provide more points than necessary to construct the bounding manifold; they will simply be removed during the manifold construction process. Only works for convex manifolds for now}
}
\value{
\item{UniformFill}{Matrix containing uniformly randomly distributed points in the supplied manifold. Column names are transferred from inputranges}
}
\description{
This function uses a hyper-rectangular proposal region and a constraining set of hull points to fill the convex hull with a uniform distribution of points
}
