% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrameProcessing.R
\name{customcontrsum}
\alias{customcontrsum}
\title{Change factor coding to contr.sum}
\usage{
customcontrsum(factorin, speclevels)
}
\arguments{
\item{factorin}{vector of factor values}

\item{speclevels}{vector of factor level names in desired order}
}
\value{
vector as a factor with contrast sum encoding that respects the order in speclevels
}
\description{
Converts a factor contrast to contr.sum while preserving factor order and applying appropriate column names.
  Changes factor codings to contr.sum without changing global factor settings for R.
}
