% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defficiencyfuncs.R
\name{d_effchoiceupdate}
\alias{d_effchoiceupdate}
\title{D-efficiency update function for single question for multinomial logit model}
\usage{
d_effchoiceupdate(CurrentMatrix, paramestimates, info_mat = 0)
}
\arguments{
\item{CurrentMatrix}{subsection of model matrix for one question only. Should include all alternatives for that questionas created by the model.matrix function with all continuous parameters centered and standardized
and appropriate factor numerical coding (generated using contr.sum for factors) as created by the model.matrix function.
Data.frame formats not accepted
Opt out choices, if included, should be coded as a row of all zeroes for every parameter}

\item{paramestimates}{standardized effect estimates for each effect (column) of model matrix corresponding to standardized model matrix.}

\item{info_mat}{the information matrix for all other questions in the model matrix}
}
\value{
numeric value for d-efficiency
}
\description{
Fast d-efficiency update function for use during search algorithms to reduce search time. Requires supplied information matrix for all questions not included
}
