% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defficiencyfuncs.R
\name{d_efficiency}
\alias{d_efficiency}
\title{D-efficiency calculation}
\usage{
d_efficiency(CurrentMatrix, det_ref, input_formula, Input_range)
}
\arguments{
\item{CurrentMatrix}{either a matrix or data.frame of base variables only (ex: Col for A, B, C) that will be converted to a model matrix using input_formula}

\item{det_ref}{reference determinant for calculation of d_efficiency. Should be max attainable determinant of info matrix}

\item{input_formula}{one-sided formula (inputs only) for model to calculate d efficiency from. Format = ~ A + B + A:B + I(A/B^2), etc
required if CurrentMatrix is a data frame of base variables that needs to be converted to a model matrix}

\item{Input_range}{matrix or data frame listing the range of columns used for standardization. Column names must match input_formula term names.
format is matrix or data frame with column in Input_range matching each column name in CurrentMatrix and minimum value in first row and maximum value in second row.}
}
\value{
vector(d_eff,det_calc) where the first element is the d-efficiency and the second element is the determinanat of info matrix created using CurrentMatrix and input_formula
}
\description{
Calculates determinant ratio based on supplied det_ref. Will convert a base data.frame/matrix to a model matrix and standardize based on supplied input_formula and Input_range.
}
