% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defficiencyfuncs.R
\name{d_efficiencysimple}
\alias{d_efficiencysimple}
\title{Linear Model D-Optimality. Review for deprecation and removal.}
\usage{
d_efficiencysimple(CurrentMatrix, returncov = FALSE)
}
\arguments{
\item{CurrentMatrix}{matrix - A model matrix with appropriate factor numerical coding as created by the model.matrix function. Data.frame formats not accepted}

\item{returncov}{logical - Whether the variance-covariance matrix should be returned (requires additional calculation time)}
}
\value{
Either the D-Optimality of the model if returncov is FALSE or a list with elements d_eff (D-Optimality) and vcov (covariance matrix) if returncov is TRUE
}
\description{
Calculates determinant of supplied model and raises it to the power of 1 / ncol(ModelMatrix) and divides the result by the number of rows in ModelMatrix.
}
