% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defficiencyfuncs.R
\name{d_efflinearupdate}
\alias{d_efflinearupdate}
\title{Linear model D-Optimality update function. Review for deprecation and removal.}
\usage{
d_efflinearupdate(CurrentMatrix)
}
\arguments{
\item{CurrentMatrix}{matrix - A model matrix with appropriate factor numerical coding as created by the model.matrix function. Data.frame formats not accepted}
}
\value{
The D-Optimality of the model
}
\description{
Calculates D-Optimality of the supplied model matrix by calculating the determinant and raising it to the power of 1 / ncol(ModelMatrix) then dividing the result by the number of rows in ModelMatrix.
}
