% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrameProcessing.R
\name{designmommat}
\alias{designmommat}
\title{Calculate moment matrix of design space}
\usage{
designmommat(inputranges, modelformula, constrainnodes = NA, n = 1000 *
  ncol(attributes(stats::terms(modelformula))$factors))
}
\arguments{
\item{inputranges}{matrix - A matrix with named columns where first row is minimum and second row is maximum - range of base input variables that will be used as ranges to generate random points for design space filling (including geometric combination variables)}

\item{modelformula}{formula - The formula of the model to use for calculating moment matrix. Terms must match column names of inputranges otherwise model matrix cannot be calculated appropriately}

\item{constrainnodes}{matrix - A matrix with nodes of constraining manifold (hull) within rectangular inputranges space.
Should be a matrix of nodes as produced by the as produced by \code{\link[geometry]{convhulln}}.
It is ok to provide more points than necessary to construct the bounding manifold; they will simply be removed during the manifold construction process. Only works for convex manifolds for now}

\item{n}{integer - The number of points to use to numerically approximate moment matrix integral. It is suggested that at least 1000 points per factor in model formula be used}
}
\value{
\item{MomentMatrix}{The moment matrix}
}
\description{
Approximates the moment matrix of the design space calculated using numeric approximation of moment matrix based on uniform distribution of random points specified by the inputs.
  This moment matrix estimate is typically used to calculate I-Optimality for I-Optimal design creation
}
