% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tournament.R
\name{gentourneybracket}
\alias{gentourneybracket}
\title{Create choice tournament bracket structure}
\usage{
gentourneybracket(altvect, ChoiceTournamentAlts = 2)
}
\arguments{
\item{altvect}{vector - vector defining which row of the design belongs to each question set for the starting questions (not including tournament bracket that will be generated by this function)}

\item{ChoiceTournamentAlts}{integer, default 2 - Maximum number of alternatives that should be used for each tournament matchup}
}
\value{
\item{matchupframe}{Matrix that details all possible matchups based on tournament bracket structure.}
  \item{tournamentframe}{Matrix showing the tournament bracket structure.}
}
\description{
Creates a full tournament bracket from the starting matchups described in altvect with no matchup containing more than ChoiceTournamentAlts profiles.
  Tournament will always be complete (one winning profile remaining) but may include one or more choice sets with fewer than ChoiceTournamentAlts in a match if necessary.
}
