% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tournament.R
\name{getchoicemodinfoprobs}
\alias{getchoicemodinfoprobs}
\title{Calculate performance metrics for choice tournament designs}
\usage{
getchoicemodinfoprobs(modmat, paramestimates, matchupframe,
  usepriorwinprobs = TRUE)
}
\arguments{
\item{modmat}{matrix - Model matrix of design to evaluate (produced using model.matrix(formula, design matrix))}

\item{paramestimates}{vector - Coefficient estimates to use for calculating probabilities for the formula passed in modform.
Position must match coefficient column positions in modmat.}

\item{matchupframe}{matrix - matrix returned by \code{\link{gentourneybracket}} that ties each row in designmat to the corresponding choice set, including all potential tournament matchups}

\item{usepriorwinprobs}{logical, default TRUE - whether win probabilities should be calculated using prior parameter estimates or not. If FALSE, assumes each alternative has an equal chance of winning each round (may provide better protection against prior misspecification).
Otherwise, uses prior estimates to calculate win probabilities for each choice set}
}
\value{
\item{setproplist}{Information matrix for each choice set. RENAME THIS OUTPUT IN THE FUTURE TO BETTER MATCH CONTENTS.}
  \item{winprobvect}{Vector of win probabilities for each profile in matchupframe}
  \item{occurprobvect}{A vector of the probability of each choice set occurring.}
  \item{info_mat}{Information matrix of the tournament design}
}
\description{
Calculates metrics for choice tournament designs based on the prior parameter estimates and the tournament bracket.
  The information matrix is the sum of the information matrix for each possible choice set in the tournament multiplied by its probability of occurence.
}
