% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defficiencyfuncs.R
\name{matchupprobs}
\alias{matchupprobs}
\title{Matchup probability for choice tournament}
\usage{
matchupprobs(matchupframe, winprobs, occurprobs = NA, updatevect = NA)
}
\arguments{
\item{matchupframe}{data frame with the following columns:
matrixrowid - unique number corresponding to the alternative ("team") as it progresses through the tournament.
questionid - number of the matchup ("game").
uniquesetid - unique number assigned to each potential matchup. There will be multiple potential matchups for each possible questionid
sourcequestion - the previous questionid ("game") that the alternative ("team") must have won to arrive in this questionid ("game"). Should be 0 for the first questionid ("game") for each alternative ("team")}

\item{winprobs}{vector of win probabilities for each alternative in matchupframe. Must have length equal to number of rows in matchupframe}

\item{occurprobs}{(optional) vector of pre-existing chance of each matchup occurring. Should be 1 for any questionid that does not have a sourcequestion in matchupframe. If not passed, the entire occurrence probability vector will be generated from scratch.}

\item{updatevect}{(optional) vector of questionids in matchupframe that should be updated based on the supplied winprobs vector and previous occurprobs vector. If not passed, all occurrence probabilities will be updated.}
}
\value{
vector of probability of each matchup occurring
}
\description{
Calculates the probability of occurrence for every possible matchup in a choice tournament bracket
}
