% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizers.R
\name{optimizefn}
\alias{optimizefn}
\title{Optimizer update function to exchange point and recalculate objetive function}
\usage{
optimizefn(x, index, roworcol, designmat, base_input_range,
  formulacollection, scalemat = TRUE,
  weight = rep(1/length(formulacollection), times =
  length(formulacollection)))
}
\arguments{
\item{x}{vector of numerics - new values for the row, column, or coordinate of design matrix being modified.
Length must match destination length of target row, column, or coordinate of \code{designmat}}

\item{index}{integer - position of row or column being modified}

\item{roworcol}{character - whether to replace row, column, or coordinate.
Uses "row", "column", or "coordinate" as argument choices}

\item{designmat}{whole design matrix in raw variable terms}

\item{base_input_range}{input type - 2 row matrix - matrix with variable minimums and maximums for base variables (variables common to all formulas in formulacollection). First row is minimum, second row is maximum. Names of colums should equal variable names used in formulas}

\item{formulacollection}{collection of formula objects used to construct scaled model matrices. Should include the following at a minimum for linear models. Additional arguments can be included for more complex operations
\item{inputranges}{matrix - A matrix with one column per variable in inputmat and minimum and maximum values in row 1 and 2, respectively}
\item{baseformula}{formula - no-intercept formula of base algebraic combination variables}
\item{range}{matrix - matrix with one column per variable in baseformula and minimum and maximum values in row 1 and 2, respectively}
\item{termsalt}{vector, optional - names of alternate terms to replace terms in baseformula with so that the model matrix function will work. If not provided, returns scaled model matrix for baseformula}
\item{retermedformula}{formula, optional - full model formula with terms replaced by altterms elements. If not provided, returns scaled model matrix for baseformula}}

\item{scalemat}{logical, default TRUE - whether the base variable matrix for each model should be scaled from -1 to 1 or not}

\item{weight}{vector - weights to apply to each formula's objective function for overall objective function}
}
\value{
The objective function value for the updated design with the new row, column, or coordinate value(s)
}
\description{
Recalculates an objective function using the new row, column, or coordinate. Used for design optimization.
}
