% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAnalysis.R
\name{ploteffs}
\alias{ploteffs}
\title{Assess discrete chioce design metrics over a range of true parameter values}
\usage{
ploteffs(designlist, typevec, trueframe, design_params, refindex = NULL,
  linetype = rep("solid", length(designlist)), prevplots = NULL,
  returntype = "plotly", loess_smooth_plots = FALSE)
}
\arguments{
\item{designlist}{A list of discrete choice designs created by \code{\link{optimizemodellist}}}

\item{typevec}{A vector of strings describing the discrete choice design types. Options are:
\describe{
  \item{"Regular"}{A standard discrete choice design without a tournament.}
  \item{"Tournament"}{A fixed bracket choice tournament design where the same bracket assignments are used for each tournament replicate. (Includes forward-looking designs)}
  \item{"RandTournament"}{A random bracket choice tournament design where bracket assignments are randomized after each tournament completion.}
  }}

\item{trueframe}{The model matrix of true parameter values to iterate across where each row is a combination of true parameter values that should have a result returned.}

\item{design_params}{The vector of assumed parameter values to calculate distance from for each plot. Typically the vector of assumed parameter means or modes used to create the designs being assessed.}

\item{refindex}{Integer giving reference design position in designlist if design efficiencies should be returned with respec to that particular design.}

\item{linetype}{Type of line to use for each plot (only used if returntype is "plotly")}

\item{prevplots}{The return object created by a previous call to ploteffs if additional results should be added to that object.}

\item{returntype}{A string specifying the type of results to return. Options are:
\describe{
  \item{"plotly"}{Returns a list of plotly graph objects where each design in designlist will have a corresponding line on the plot.}
  \item{Any other argument}{Returns a list of numeric results where each design in designlist will have a corresponding entry in the list for each metric type.}
  }}

\item{loess_smooth_plots}{If TRUE, applies loess smoothing to efficiency results}
}
\value{
Returns a list of plots or a list of results with the following entries
  \item{x}{The x-values used for the plot}
  \item{seB1}{The standard error of the first coefficient in the design.}
  \item{DOpt}{The D-Optimality or D-Efficiency (if refindex is provided) of each design}
  \item{IOpt}{The I-Optimality or I-Efficiency (if refindex is provided) of each design}
  \item{OpCR}{The size of a confidence region centered at the true optimum in the design space that is 1 standard error in radius or the confidence region size relative to the reference design confidence region size if refindex is provided}
  \item{AOpt}{The A-Optimality or A-Efficiency (if refindex is provided) of each design}
  \item{ProbVar}{The average probability variance of each choice set or the ratio of the average probability variance to the reference design average probability variance if refindex is provided}
  \item{OptVar}{The prediction variance at the true optimum in the design space or the prediction variance relative to the reference design prediction variance if refindex is provided}
}
\description{
Plotting function for model efficiencies across a range of parameters (hard-coded) with optional overlay.
}
