% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignAnalysis.R
\name{plotlycovmat}
\alias{plotlycovmat}
\title{Create a surface plot of prediction variance against two variables}
\usage{
plotlycovmat(covmat, model_formula, xvar, yvar, xlims = c(-1, 1),
  ylims = c(-1, 1), steps = 100, prevplot = NULL)
}
\arguments{
\item{covmat}{The covariance matrix for the design}

\item{model_formula}{The formula for the model (must match formula used to create covariance matrix)}

\item{xvar}{The variable name to show on the x-axis of the plot. Must match base term in model_formula}

\item{yvar}{The variable name to show on the y-axis of the plot. Must match base term in model_formula}

\item{xlims}{A vector giving the limits of the plot for xvar}

\item{ylims}{A vector giving the limits of the plot for yvar}

\item{steps}{An integer specifying the number of grid-points along each axis to use to create the plot}

\item{prevplot}{A previous plotly plot object if this surface should be added to an existing surface created by this function.}
}
\value{
A plotly surface plot of the prediction variance against xvar and yvar.
}
\description{
This function creates a surface plot of prediction variance against two predictor variables.
  Requires the plotly library to function.
}
