% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerFuncs.R
\name{vcovpower}
\alias{vcovpower}
\title{Power calculation for coefficients}
\usage{
vcovpower(vcovmat, detectdiff, test_alpha = 0.05, test_beta = 0.2)
}
\arguments{
\item{vcovmat}{variance-covariance matrix for model to estimate}

\item{detectdiff}{vector containing minmum detectable difference for each effect (column) of model matrix sized corresponding to standardized model matrix.}

\item{test_alpha}{numeric value or vector specifying allowable alpha error. If different alpha values are to be used for each estimated parameter, vector order must match}

\item{test_beta}{numeric value or vector specifying allowable beta error. If different beta values are to be used for each estimated parameter}
}
\value{
data frame with the following:
  first column containing parameter name
  second column containing difference to detect
  third column containing minimum sample size required to reach supplied power value
  fourth column containing alpha value used
  fifth column containing beta value used
}
\description{
Calculates power to detect each supplied coefficient difference based on supplied alpha and beta error.
}
