% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFrameProcessing.R
\name{Xmat}
\alias{Xmat}
\title{Calculates scaled model matrix}
\usage{
Xmat(inputmat, inputranges, baseformula, baseformrange, altterms = NULL,
  fullformulareterm = NULL)
}
\arguments{
\item{inputmat}{matrix - design to evaluate in terms of variables in inputranges}

\item{inputranges}{matrix - matrix with one column per variable in inputmat and minimum and maximum values in row 1 and 2, respectively}

\item{baseformula}{formula - no-intercept formula of base algebraic combination variables}

\item{baseformrange}{matrix - matrix with one column per variable in baseformula and minimum and maximum values in row 1 and 2, respectively}

\item{altterms}{vector, optional - names of alternate terms to replace terms in baseformula with so that the model matrix function will work. If not provided, returns scaled model matrix for baseformula}

\item{fullformulareterm}{formula, optional - full model formula with terms replaced by altterms elements. If not provided, returns scaled model matrix for baseformula}
}
\value{
Scaled model matrix for the design
}
\description{
Function to return properly scaled model matrix based off of input matrix of base variables.
  This function is compatible with constructed variables as well (e.g. I(A/B)).
}
