% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Defficiencyfuncs.R
\name{d_effchoice}
\alias{d_effchoice}
\title{D-Error for multinomial logit model}
\usage{
d_effchoice(CurrentMatrix, altvect, paramestimates = NULL,
  returninfomat = FALSE)
}
\arguments{
\item{CurrentMatrix}{model matrix as created by model.matrix function with all continuous parameters centered and standardized
and appropriate factor numerical coding (generated using contr.sum for factors) as created by the model.matrix function.
Data.frame formats not accepted
Opt out choices, if included, should be coded as a row of all zeroes for every parameter}

\item{altvect}{vector with integer corresponding to each row that indicates which choice set it is a member of}

\item{paramestimates}{estimates for each effect (column) of model matrix sized corresponding to standardized model matrix.
If not supplied, parameter estimates will be assumed equal to zero for all parameters}

\item{returninfomat}{whether to return the covariance matrix as well. Default is FALSE}
}
\value{
list containing:
  d_error - d-error of supplied design with respect to parameter estimates
  covmat - covariance matrix
}
\description{
Calculates d-error of supplied model matrix. Will also calculate probability centered d-error if vector of parameter estimates is supplied
}
